@@ -1,6 +1,6 @@
 {
   "name": "react-reflex",
-  "version": "1.0.13",
+  "version": "1.0.14",
   "description": "Flex layout component for advanced React web applications",
   "main": "dist/lib/index.min.js",
   "scripts": {
@@ -45,7 +45,7 @@
     "react-measure": "^1.4.5",
     "sass-loader": "^2.0.1",
     "style-loader": "^0.12.4",
-    "webpack": "^2.1.0-beta.28"
+    "webpack": "^2.2.1"
   },
   "dependencies": {
     "babel-runtime": "^6.23.0",
